Фиксы

+++ Lexplosion.Core/Logic/Management/Instances/InstanceClient.cs
@@ -230,7 +230,7 @@ namespace Lexplosion.Logic.Management.Instances
 			}
 		}
 
-		public int DownloadCounts { get; internal set; } = -1;
+		public long DownloadCounts { get; internal set; } = -1;
 		public bool HasDownloadCounts { get => DownloadCounts > -1; }
 
 		private string _websiteUrl = null;
